Added a new msbuild command to use the Microsoft Build Tools 2015
Added a new msbuild command to use the Microsoft Build Tools 2015 instead of the default MsBuild.exe provided by the Framework. This is because the new features in .NET Framework 4.6 and above aren't supported by the default msbuild.
To use this is required to install in the agents the Microsoft Build Tools 2015.